Belt-Drive Gearbox Prototype 1
Very difficult to assemble, lacks a working, easily adjustable tensioning mechanism and does not account for the poor tolerance in cheap bearings. In the next design i will seek to improve this by separating the motor input gear from a structural shaft, redesigning a tensioning mechanism that can be easily adjusted with screws. 


1. Loosen tolerance on printed rod holes.
2. Remove tensioner wheels, tension the structural shaft directly, creating space between the gearsets. (Use rods for alignment if needed.)
3. Flip the motor to make the design more compact.
4. Keep the overall structure rigid.